Meal Tabs Nav bar
Started working on the meals tabs nav bar. I also fixed the search button so now it changes color when you hover over it. I still need to fix the vertical centering of the search box. Also need to finish the nav bar. This is just the start of the nav bar.